[xdl] [xde] [exponent-sdk] Stream logs to xdl
Summary: XDL sends a new field `logUrl` with the manifest. When the field is available, exponent-sdk overrides console.* to send logs to that url in addition to the original console.*. Logs are added to a queue and then sent in batches to avoid xdl receiving them out of order. XDL pipes this through bunyan so that exp automatically sees them too. XDE intercepts device logs and shows them in a new tab.

Test Plan: Test with new starter project

Reviewers: nikki, ben

Subscribers: ide

Differential Revision: https://phabricator.sixfivezero.net/D2992