[Assets] Use `assetPlugin` parameter in SDK 10.1.0+
Summary:
This is a new parameter supported by the React Native packager so we don't need to put Exponent-specific code into React Native. Instead we ship a small plugin with the `exponent` package and then load it from XDL when a project says it's using a new SDK version.

You already need `exponent` to use assets so I don't think this changes any requirements.

In SDK 12 I am planning to drop the old `includeAssetFileHashes` parameter from the packager (we'll leave the code in XDL to support old projects) but people need this new XDL code for SDK 12.

Test Plan:
Installed a local copy of exp-latest React Native (SDK 10.1.0 candidate with assetPlugin support) in the new project template app. Installed a local copy of exponent-sdk in it as well so that the tools are available. Modified exp.json to say sdkVersion 10.1.0.

Loaded the manifest in a browser and saw that the bundle URL included `assetPlugin` and that loading the bundle URL produced a JS bundle with fileHashes.

```
__d(877 /* @exponent/samples/assets/images/slack-icon.png */, function(global, require, module, exports) {module.exports = require(168 /* react-native/Libraries/Image/AssetRegistry */).registerAsset({"__packager_asset":true,"httpServerLocation":"/assets/node_modules/@exponent/samples/assets/images","width":40,"height":40,"scales":[2,3],"hash":"7ccbd8969a57fa489718a5c960efcab8","name":"slack-icon","type":"png","fileHashes":["af47141a24eb8e5751dfd034159479bd","ad107d1ca027259ca50ce0f432293e6c"]});
}, "@exponent/samples/assets/images/slack-icon.png");
```

Reviewers: nikki, jesse

Differential Revision: https://phabricator.sixfivezero.net/D3035