[XDL][Tests] Add dtrace-provider mock (returns null) to clean logspew
Bunyan requires dtrace-provider in a soft way -- it doesn't actually need it but optionally sends events to dtrace probes if it can require dtrace-provider. Bunyan explicitly supports null mocks for dtrace-provider so that Bunyan can be browserified, for example, so it's OK to define a null mock for it.

The tests now don't spew out console.error noise, making them easier to sift through.