Suppress verbose extraction output in `exp init`
Summary:
exp init would spit out a long list of files that it had extracted when
you ran it. This wasn't that helpful.
This change just removes the 'v' (verbose) option from the `tar` command,
so errors should still be surfaced.

I also looked at leaving verbose output on but only surfacing stderr while
suppressing stdout, but it turns out that `tar xzvf` logs to stderr, so
that approach is not an option.

Test Plan:
    ccheever@Charless-MacBook-Air:~/universe/dev/xdl$nesh -y
    Node v6.8.1
    Type .help for more information
    nesh*> .require .
    xdl = require("/Users/ccheever/universe/dev/xdl")
    nesh*> yield xdl.Exp.createNewExpAsync('blank', '/Users/ccheever/tmp/test-exp-proj/asd' + Math.random(), {}, {name: 'trouble'})
    '/Users/ccheever/tmp/test-exp-proj/asd0.189655667206728/trouble'
    nesh*>

Reviewers: jesse

Reviewed By: jesse

Differential Revision: https://phabricator.sixfivezero.net/D3047