[XDL] Remove dtrace-provider upon postinstall
dtrace-provider is optional and causes problems because it's a native module, but we share XDL across Electron and Node which have different native ABIs. The real solution is to do something like my "yarn knit" proposal but that's not even being worked on so for now just crudely delete dtrace-provider. It's safe, I feel this specific time it's OK to do so.