[exp] Add "expert auth" mode to allow standalone builds without any apple login. (#1681)
In this mode, users will not enter their apple id or password, and will just upload certs and their provisioning profile to the build service. It will probably be really error prone, and I wouldn't think this is a good idea if we weren't about to ship user-visible build logs.

Also, fixes a bunch of flow errors in the builder classes from using an & type instead of an | type for Credentials.