[ios][standalone] Add a --testEnvironment option to the "configure" script
Makes `gulp ios-shell-app --action configure` take a new option called `--testEnvironment` which can be "local", "ci", or "none" (the default). This sets the "testEnvironment" key in EXShell.plist, which the fully configured iOS standalone app uses at runtime to (a) decide whether to register EXTest and (b) to tell EXTest what kind of testing environment we're in (e.g. CI).

Tested by running the configure script with a published copy of test suite and `--testEnvironment ci`. Used a console.log statement to ensure that the script set the "testEnvironment" field of the plist object. Then installed and ran the generated .app on the iOS simulator and saw that EXTest logged the results:

```
$ xcrun simctl spawn booted log stream --predicate '(subsystem == "host.exp.Exponent") && (category == "test")'
Filtering the log data using "subsystem == "host.exp.Exponent" AND category == "test""
Timestamp                       Thread     Type        Activity             PID    TTL
2018-02-22 19:00:36.267708-0800 0x35d836   Default     0x0                  73640  ExpoKitApp: [host.exp.Exponent:test] [TEST-SUITE-END] {"failed":0}
```