[ci] Use UNVERSIONED when building an iOS build in CI (fix "glog" error)
This changes CI to use UNVERSIONED instead of the latest SDK version in CI. This affects the base iOS app that is used for Turtle and E2E iOS tests.

The intent here is to configure the app to use the latest version of the code. Currently the rest of the build process in CI is failing because it's trying to use the unversioned copy of RN (SDK 26 / RN 0.54) but configuring the Podfile for SDK 25 / RN 0.52, since that's the latest version returned from www's version endpoint. This causes the "glog" (SDK 26) vs "GLog" (SDK 25) error we're seeing in CI.

This commit makes it so we use `UNVERSIONED` in CI. Since `UNVERSIONED` is not in the www versions response, we skip the code path that uses it. This ends up working because data from the versions response is just needed for "user" (vs. "service") builds.

Test Plan: See if we can build the iOS base apps in CI.