[json-file] Remove file locking
Locks were added to @expo/json-file at one point, but the version with
locks has not been used (9e6e7e0f258) and in my tests was significantly
slower than the previous versions. The locking mechanism also adds a lot
of complexity and slows down the operations a lot. It would require more
testing to be reliable.