[android][fcm] Enable FCM in the Expo client
We are moving to FCM for Android notifications because Google has deprecated GCM and we have deliverability issues with GCM/SNS. This commit adds the FCM services config to AndroidManifest.xml for all apps and then sets the `FCM_ENABLED` to true by default. For standalone apps that don't have a `google-services.json` file, we set `FCM_ENABLED = false`.

Tested by building the Expo client locally and running the push notification test experience. Verified in the Postgres DB's `push_devices` table that the client registered using FCM, and sent a test notification with the notification dashboard to test notifications end-to-end.

Detached a project and verified that `FCM_ENABLED = false` was in the project (which didn't have FCM set up for standalone).