[podfile] Add RCT_ENABLE_INSPECTOR=0 to the generated Podfile
The "inspector" refers to the Nuclide inspector that RN tries to connect to on port 8082. This doesn't work with Expo CLI and causes errors when we load a published copy of dev Home from CloudFront and RN tries to connect to `<cloudfront>:8082/<inspector URL>`. This makes the Podfile configure the build so the inspector is disabled and we don't make network requests to unsupported endpoints that will fail.