[xdl] Show a warning when using exp.json
The configuration code still assumes the config can be found in either
app.json (the current default everywhere) or exp.json (a legacy option).

This commit adds a warning that is shown if app.json doesn't exist but
exp.json does.

This is the first step towards removing the support for exp.json
altogether and only using app.json.