Show a note about eject when running `detach` (#384)
exp (the predecessor of Expo CLI) had a command named `detach` that would eject your project to ExpoKit. In Expo CLI this functionality is a part of the `expo eject` (which lets you choose `ExpoKit` when ejecting).

This PR adds a more helpful error message when running the `detach` command that doesn't exist.