Add image optimization functionality (#555)
* Add optimize command and read in bundled assets

* Add sharp to dependencies and compress images

* Use assets.json in .expo folder to store which assets have been modified

* Store assets by their checksum in .expo-shared and add --save option

This commit switches to storing assets.json in a new .expo-shared folder
rather than .expo so that it can be checked into git and state can be shared
amongst multiple people . It also now keeps track of files using the SHA256
checksum of their contents rather than their inodes or names which allows it
to work crossplatform and accross different machines. Finally, I added a --save
option to save the original assets and switched to overwriting them and
not saving the unoptimized versions by default.

* Delete unused hash values in assets.json and log when no assets are optimized

Deleting unused hash values allows us to remove entries in assets.json for assets that were deleted or modified

* Only search for assets files in project directory

* Output KB saved and remove 'optimized' assets if larger than original

* Update function to convert number of bytes to a human readable value

* Add warning and prompt the user if --save flag is not passed in

* Prompt user to optimize assets before publishing

* Add toggle for asset optimization in dev-tools

* Add optimize functionality to dev-tools GUI

Moves core logic of optimization to xdl and imports from within expo-cli and
dev-tools. Adds a new Graph QL mutation which the dev-tools front end
calls to trigger the xdl optimizeAsync function.

* Add message about assets.json when it's created.

* Use path.join instead of string concatenation

* Move helper functions into a new AssetUtils module

* Add --include and --exclude options to filter files

* Clarify that --include and --exclude patterns are relative to proj root

* Make logs less colorful

* Update prompt messages to be more clear

* .expo -> .orig for backups

* Remove backup prompt/warning when there are no assets to optimize

* Update prompt message to backup original files