Update publish script to handle new packages we haven't published
The publish script didn't handle the case where a package wasn't published before, causing `npm view <pkg>` to fail. This commit handles the error and also marks the package as 🆕.

Test Plan: Added `--dry-run` to the `npm publish` part of the script and ran the publish script and verified that `@expo/babel-preset-cli` was printed as new and that the script didn't fail.