Add blank-ts to set of templates you can add
I made expo-template-blank-typescript and this adds it as an option to expo-cli
This won't actually work until https://github.com/expo/universe/pull/3555 is merged, which is the
commit that adds the template.

I chose to be consistent with all the other text in the options and use all lowercase (typescript instead of
TypeScript).

I chose to add the template in the managed workflow section under the blank template which it is almost
exactly the same as.

In the future, we may want to make TypeScript support a separate question rather than just maitaining
a bunch of TS versions of templates