refactor(xdl): convert xdl.js to TypeScript and emit declarations (#801)
While we are converting XDL to TypeScript we temporarily use Babel to build both TypeScript and Flow sources (see #675). However, it would be useful to be able already to import XDL modules from TypeScript and make use of the types, but the current setup doesn't output the .d.ts declaration files.

This PR adds the --emitDeclarationOnly flag to the type checking command so that .d.ts files are generated in the build folder, to be consumed by other packages. It also converts the xdl.js public API entry point file to TypeScript.

Because some of the modules are yet to be converted, xdl.ts has some @ts-ignore directives to prevent warnings from those modules. I have also added the .js prefix to the import statements for those modules, as a reminder to update xdl.ts and remove the corresponding @ts-ignore, whenever a new module is converted.