BREAKING CHANGE: make supported Node.js versions stricter and improve the error message
We support Node.js maintenance LTS, active LTS and current release versions.
Previously the supported version range was defined as 8.x.x || >=10.0.0,
but this included the 11.x.x versions that reached enf-of-life on June 1st 2019.

Update the supported versions to include the exact ranges we support and also
make the error message clearly tell which versions are supported.

Note: the version ranges purposefully use the normal forms such as '>=8.9.0 <9.0.0'
instead of '^8.9.0' to make it easy to understand even without knowledge of the
intricacies of semver.