fix: slow CLI startup time performance regression
When changing the build configurations to add support for TypeScript, the lazy loading of modules that was previously done using the `babel-transform-inline-imports-commonjs` package was removed from `@expo/xdl` and `expo-cli`:
- https://github.com/expo/expo-cli/commit/b45e9dbf724bc077b7fbaf5041961028c4ddf470#diff-77b4343b244f7776c55c7c643d13efe6
- https://github.com/expo/expo-cli/commit/e3c7af725987e17bac74d0e806c67b0691ca626b#diff-b80134f5b17e5b157842c17b91ec70f4

This affected startup time of the CLI negatively, causing every command to take 3-5 seconds to start.

This commit restores lazy loading of modules, now using the `lazy` option of `@babel/plugin-transform-modules-commonjs`.