feat(xdl): Allow ejecting just one platform if the other one is not released yet (#1293)
# Why

Right now it's not possible to run `expo eject` if ExpoKit for one of the platforms is not released yet. It would be more convenient to allow such thing as sometimes the release process of these two platforms is at different stages.

# How

Made small changes in a few conditions.

# Test plan

Tested by running `EXPO_STAGING=1 expo eject` on SDK36 project while only iOS ExpoKit is released.