fix: update XDL caches before publishing (#1675)
The XDL package includes a cache, which is used a fallback for offline mode. The cache was previously updated automatically as a part of the each release, updating it stopped at some point.

This commit updates the cache to include the latest SDK versions and adds an npm prepack script that will be run before XDL is compressed into a tarball to be published to npm.