[ts] Use require() to read package.json instead of import for TypeScript (#1667)
By default, Node does not allow for a package to import itself. This might work in scenarios such as a Yarn monorepo where all linked workspaces are linked under the root directory's `node_modules` folder so searching up the directory hierarchy leads to `<root>/node_modules/<workspace>`. This breaks in reality when a package is taken outside of the monorepo. This notably happens with expo-cli since it is most likely to be installed as a top-level program and not as a dependency.

If we import package.json, TypeScript thinks the package directory is the root directory and emits `build/{package.json,src}`. We don't want that -- we want package.json to be left alone and for the compiled contents of `src` to be emitted into `build`.

Test plan: `rm -fr build tsconfig.tsbuildinfo; yarn tsc` and check that `build` doesn't contain package.json. Start with `node bin/expo.js`.