Fix app.json parse errors not reported correctly (#1709)
When `app.json` contains invalid json it would fall-though the config loading process and result in an empty JSON object. No error message is then printed pointing to the location in the JSON file that contains the syntax error.
Instead, schema validation errors such as ‘no slug` or ‘no name` would be printed which are confusing.