Convert commands to use dynamic imports
This was an experiment to try speeding up the CLI built with ncc,
however, it seems that this doesn't really help with the slow
startup with ncc builds. I assume the problem is that majority
of the time is spent parsing the 36M(!) bundle and dynamic imports
only help with the execution time.