Update "googleMobileAdsAutoInit" to be optional
Following the information provided both in this file and on the docs (found here: https://docs.expo.io/workflow/configuration/#android) - "The default in Expo (Client and in standalone apps) is `false`". Drawing from this I have concluded that this value is not actually required, hence updating the type declaration to reflect this.