feat(expo-cli): Add 'validate-dependencies' command
Adds `expo validate-dependencies`.

This command will terminate with a non-zero exit code if a dependency version is
incompatible with the projects SDK version.

It is designed to be used with automated dependency managers such as Dependabot
as part of the CI process to verify if native dependencies will be compatible
with the Expo SDK version.