[expo-cli] Support absolute path in credentials.json for gradle
Why
===
We handle both absolute and relative paths in `credentials.json` in other places, but it wasn't handled in the gradle script. This commit fixes it.

Test plan
=========
Try both absolute and relative paths in `crdentials.json` for the `keystorePath` in the turtle example project and make sure that the release build passes with `./gradlew assembleRelease`.