[expo-cli] Add `releaseChannel` field to the profile in `eas.json`
Why
===
We want to be able to specify the release channel when building our app so that subsequent updates `expo-updates` can be published separately for different profiles.

How
===
I've added the `releaseChannel` field to the `jobs` and `metadata` objects which are sent to the www server when running `expo eas:build`.

Test plan
=========
I verified that the `releaseChannel` is respected by adding a new `releaseChannel` field under the profile in the `eas.json` of the turtle example app, and then running the build to make sure that tis field is received by the turtle worker as well.