[json-file] Remove lodash from @expo/json-file
`@expo/json-file` previously used `lodash/get` and `lodash/set` for the
`getAsync` and `setAsync` methods.

In theory, it has been possible to pass something like `'some.key.here'`
to these methods, which will no longer work after this change. However
this behavior was 1) undocumented 2) unsupported by the TS types and
3) not used anywhere in the `expo-cli` code base. Anyway, it's probably
safest to make a major version bump for `@expo/json` CLI because of this.