[Interstitial page] Ensure that development build is installed (#4144)
# Why

Close ENG-2211.

# How

Before opening a deep link to the interstitial page, we need to check if everything is installed. So I've added a logic that will install the `Expo Go` and check if the development build is available. If not we will fall back to the expo go. 

# Test Plan

Test scenarios:
- app without dev-client 
- check if the expo go will be installed if not present 
- check what happens if the development build isn't present 
- check what happens if both development build and expo go isn't available