[Interstitial page] Try to detect if a deep link succeed (#4209)
# Why

Currently, if someone scans the QR code, we can't detect if the device can handle a deep link to `Expo Go` or `Development Build`. So we're trying to detect if the redirection was successful. If not we display a message. 

# How

When the user clicks, we trigger a timeout. After some time, we check if the document became inactive (if so, we know that the redirection was successful). This flow gets more complicated on iOS because the user will see a modal. In that case, we need to fire the timeout after the document will be focused. 

# Test Plan

- manual tests with Android device and iOS simulator.