fix: Support scoped package names for initialising projects
The `npm pack` produces a tgz file and shows the path with the `filename` property of the command output.
Unfortunately for scoped package names, this is outputting an incorrect filename.  This is causing initialising
expo projects from a template with a scoped package name to fail.

This PR fixes this.  To manually test this, use the following command:

```
expo init RealmJsApp -t @realm/expo-template-js
```

This closes #4416