[install-expo-modules] Replace xcode with xcparse (#4554)
# Why

`simple-plist@1.3.1` is a transitive dependency of the xcode package. it has [an incompatible issue](https://github.com/wollardj/simple-plist/issues/58) with ncc. the issue is pending for a while without publishing newer version as latest dist-tag. that blocks us from updating install-expo-modules.

# How

replace xcode with the awesome xcparse. this pr introduces a `withXCParseXcodeProject` config-plugin internally.

# Test Plan

- ✅ unit test passed
- ✅ manual test on a react-native 0.68 project
- ✅ manual test on a react-native 0.69 project